+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - Add support for NVIDIA graphics drivers.
 - Add support for Intel OpenGL graphics drivers on SteamOS.
 - Add support for changing the locale as part of the first-time setup.
+- Add support for doing a system upgrade as part of the first-time setup.
 
 ## 2.2.0